Renamed settings_page.dart to settings.dart
